Move FOQUS code to separate module with proper license
